Failing test for pg_job_manager.delete_old_jobs with psycopg2-based lib
